Refactor CI Tags provider
