+++ dd-java-agent/agent-ci-visibility/src/test/groovy/datadog/trace/civisibility/ci/BitBucketInfoTest.groovy
@@ -1,6 +1,6 @@
 package datadog.trace.civisibility.ci
 
-class BitBucketInfoTest extends CITagsProviderImplTest {
+class BitBucketInfoTest extends CITagsProviderTest {
 
   @Override
   String getProviderName() {
