+++ dd-java-agent/agent-ci-visibility/src/test/groovy/datadog/trace/civisibility/ci/CITagsProviderTest.groovy
@@ -15,7 +15,7 @@ import spock.lang.Specification
 import java.nio.file.Path
 import java.nio.file.Paths
 
-abstract class CITagsProviderImplTest extends Specification {
+abstract class CITagsProviderTest extends Specification {
 
   static final CI_WORKSPACE_PATH_FOR_TESTS = "ci/ci_workspace_for_tests"
   static final GIT_FOLDER_FOR_TESTS = "git_folder_for_tests"
@@ -45,8 +45,11 @@ abstract class CITagsProviderImplTest extends Specification {
     }
 
     when:
+    CIProviderInfoFactory ciProviderInfoFactory = new CIProviderInfoFactory(GIT_FOLDER_FOR_TESTS)
+    def ciProviderInfo = ciProviderInfoFactory.createCIProviderInfo(getWorkspacePath())
+    def ciInfo = ciProviderInfo.buildCIInfo()
     def ciTagsProvider = ciTagsProvider()
-    def tags = ciTagsProvider.getCiTags(getWorkspacePath())
+    def tags = ciTagsProvider.getCiTags(ciInfo)
 
     then:
     if (isCi()) {
@@ -67,8 +70,11 @@ abstract class CITagsProviderImplTest extends Specification {
     environmentVariables.set(GitInfo.DD_GIT_COMMIT_SHA, "1234567890123456789012345678901234567890")
 
     when:
+    CIProviderInfoFactory ciProviderInfoFactory = new CIProviderInfoFactory(GIT_FOLDER_FOR_TESTS)
+    def ciProviderInfo = ciProviderInfoFactory.createCIProviderInfo(getWorkspacePath())
+    def ciInfo = ciProviderInfo.buildCIInfo()
     def ciTagsProvider = ciTagsProvider()
-    def tags = ciTagsProvider.getCiTags(getWorkspacePath())
+    def tags = ciTagsProvider.getCiTags(ciInfo)
 
     then:
     tags.get(Tags.GIT_COMMIT_SHA) == "1234567890123456789012345678901234567890"
@@ -83,8 +89,11 @@ abstract class CITagsProviderImplTest extends Specification {
     environmentVariables.set(GitInfo.DD_GIT_REPOSITORY_URL, "local supplied repo url")
 
     when:
+    CIProviderInfoFactory ciProviderInfoFactory = new CIProviderInfoFactory(GIT_FOLDER_FOR_TESTS)
+    def ciProviderInfo = ciProviderInfoFactory.createCIProviderInfo(getWorkspacePath())
+    def ciInfo = ciProviderInfo.buildCIInfo()
     def ciTagsProvider = ciTagsProvider()
-    def tags = ciTagsProvider.getCiTags(getWorkspacePath())
+    def tags = ciTagsProvider.getCiTags(ciInfo)
 
     then:
     tags.get(Tags.GIT_REPOSITORY_URL) == "local supplied repo url"
@@ -97,8 +106,11 @@ abstract class CITagsProviderImplTest extends Specification {
     }
 
     when:
+    CIProviderInfoFactory ciProviderInfoFactory = new CIProviderInfoFactory(GIT_FOLDER_FOR_TESTS)
+    def ciProviderInfo = ciProviderInfoFactory.createCIProviderInfo(getWorkspacePath())
+    def ciInfo = ciProviderInfo.buildCIInfo()
     def ciTagsProvider = ciTagsProvider()
-    def tags = ciTagsProvider.getCiTags(getWorkspacePath())
+    def tags = ciTagsProvider.getCiTags(ciInfo)
 
     then:
     if (isWorkspaceAwareCi()) {
@@ -126,7 +138,11 @@ abstract class CITagsProviderImplTest extends Specification {
 
     then:
     if (isCi()) {
-      def tags = ciTagsProvider.getCiTags(getWorkspacePath())
+      CIProviderInfoFactory ciProviderInfoFactory = new CIProviderInfoFactory(GIT_FOLDER_FOR_TESTS)
+      def ciProviderInfo = ciProviderInfoFactory.createCIProviderInfo(getWorkspacePath())
+      def ciInfo = ciProviderInfo.buildCIInfo()
+      def tags = ciTagsProvider.getCiTags(ciInfo)
+
       tags.get(Tags.GIT_REPOSITORY_URL) == "https://some-host/some-user/some-repo.git"
       tags.get(Tags.GIT_BRANCH) == "master"
       tags.get(Tags.GIT_COMMIT_SHA) == "0000000000000000000000000000000000000000"
@@ -155,8 +171,7 @@ abstract class CITagsProviderImplTest extends Specification {
     gitInfoProvider.registerGitInfoBuilder(new UserSuppliedGitInfoBuilder())
     gitInfoProvider.registerGitInfoBuilder(new CIProviderGitInfoBuilder())
     gitInfoProvider.registerGitInfoBuilder(new CILocalGitInfoBuilder(GIT_FOLDER_FOR_TESTS))
-    CIProviderInfoFactory ciProviderInfoFactory = new CIProviderInfoFactory(GIT_FOLDER_FOR_TESTS)
-    return new CITagsProviderImpl(gitInfoProvider, ciProviderInfoFactory)
+    return new CITagsProvider(gitInfoProvider)
   }
 
   def "resolve"(workspace) {
