+++ dd-java-agent/agent-ci-visibility/src/test/groovy/datadog/trace/civisibility/ci/CircleCIInfoTest.groovy
@@ -1,6 +1,6 @@
 package datadog.trace.civisibility.ci
 
-class CircleCIInfoTest extends CITagsProviderImplTest {
+class CircleCIInfoTest extends CITagsProviderTest {
 
   @Override
   String getProviderName() {
