+++ dd-java-agent/agent-ci-visibility/src/test/groovy/datadog/trace/civisibility/ci/JenkinsInfoTest.groovy
@@ -1,6 +1,6 @@
 package datadog.trace.civisibility.ci
 
-class JenkinsInfoTest extends CITagsProviderImplTest {
+class JenkinsInfoTest extends CITagsProviderTest {
 
   @Override
   String getProviderName() {
