+++ dd-java-agent/agent-ci-visibility/src/test/groovy/datadog/trace/civisibility/ci/TeamcityInfoTest.groovy
@@ -1,6 +1,6 @@
 package datadog.trace.civisibility.ci
 
-class TeamcityInfoTest extends CITagsProviderImplTest {
+class TeamcityInfoTest extends CITagsProviderTest {
 
   @Override
   String getProviderName() {
