+++ dd-java-agent/agent-ci-visibility/src/test/groovy/datadog/trace/civisibility/ci/UnknownCIInfoTest.groovy
@@ -8,9 +8,9 @@ import datadog.trace.civisibility.git.CIProviderGitInfoBuilder
 
 import java.nio.file.Paths
 
-class UnknownCIInfoTest extends CITagsProviderImplTest {
+class UnknownCIInfoTest extends CITagsProviderTest {
 
-  def workspaceForTests = Paths.get(getClass().getClassLoader().getResource(CITagsProviderImplTest.CI_WORKSPACE_PATH_FOR_TESTS).toURI())
+  def workspaceForTests = Paths.get(getClass().getClassLoader().getResource(CITagsProviderTest.CI_WORKSPACE_PATH_FOR_TESTS).toURI())
 
   @Override
   String getProviderName() {
@@ -42,8 +42,11 @@ class UnknownCIInfoTest extends CITagsProviderImplTest {
     ]
 
     when:
+    CIProviderInfoFactory ciProviderInfoFactory = new CIProviderInfoFactory(GIT_FOLDER_FOR_TESTS)
+    def ciProviderInfo = ciProviderInfoFactory.createCIProviderInfo(workspaceForTests)
+    def ciInfo = ciProviderInfo.buildCIInfo()
     def ciTagsProvider = ciTagsProvider()
-    def ciTags = ciTagsProvider.getCiTags(workspaceForTests)
+    def ciTags = ciTagsProvider.getCiTags(ciInfo)
 
     then:
     ciTags == expectedTags
@@ -56,9 +59,11 @@ class UnknownCIInfoTest extends CITagsProviderImplTest {
     gitInfoProvider.registerGitInfoBuilder(new CIProviderGitInfoBuilder())
     gitInfoProvider.registerGitInfoBuilder(new CILocalGitInfoBuilder("this-target-folder-does-not-exist"))
     CIProviderInfoFactory ciProviderInfoFactory = new CIProviderInfoFactory("this-target-folder-does-not-exist")
-    def ciTagsProvider = new CITagsProviderImpl(gitInfoProvider, ciProviderInfoFactory)
 
-    def ciTags = ciTagsProvider.getCiTags(workspaceForTests)
+    def ciProviderInfo = ciProviderInfoFactory.createCIProviderInfo(workspaceForTests)
+    def ciInfo = ciProviderInfo.buildCIInfo()
+    def ciTagsProvider = new CITagsProvider(gitInfoProvider)
+    def ciTags = ciTagsProvider.getCiTags(ciInfo)
 
     then:
     ciTags.get("$Tags.CI_WORKSPACE_PATH") == null
