+++ dd-java-agent/agent-ci-visibility/src/test/groovy/datadog/trace/civisibility/ci/UserSuppliedCIInfoTest.groovy
@@ -1,6 +1,6 @@
 package datadog.trace.civisibility.ci
 
-class UserSuppliedCIInfoTest extends CITagsProviderImplTest {
+class UserSuppliedCIInfoTest extends CITagsProviderTest {
 
   @Override
   String getProviderName() {
